refactor(enhancement): rename catalogue to cmap
